move log to utils folder
